small fixes, remaining var type errors

+++ rvas-core_v0.1.1/src/main/java/core/backend/FileManager.java
@@ -76,7 +76,7 @@ public class FileManager {
 		);
 		
 		Files.readAllLines(killstats_user_database.toPath()).forEach(val ->
-		PlayerMeta.KillStats.put(UUID.fromString(val.split(":")[0]), Double.parseDouble(val.split(":")[1]))
+		PlayerMeta.KillStats.put(UUID.fromString(val.split(":")[0]), Integer.parseInt(val.split(":")[1]))
 		);
 	}
 }
